requirements.txt
dateutil
django-bootstrap-v5
django-crispy-forms
crispy-bootstrap5
